refactor: three cross-feature imports removed, baseline 13 -> 10 [GH-000]
The previous commit measured 13 and held the line. Three of them had a
remedy the architecture rule already prescribes, so they are gone rather
than recorded.

Query keys (payments, 2 of them). checkout imported the orders list's
cache key to invalidate it after payment; received-orders imported the
assigned list's for the same reason. A cache key is not feature-private
in the first place -- it has to be globally unique to work at all, and
invalidation is inherently one feature acting on another's data. All
nine now live in shared/domain/queryKeys.ts, where their uniqueness is
checkable by reading one file. 17 files rewired.

Audit logging (admin, 1). features/users imported useLogExport from
features/audit. Reading entries is what the audit feature is for; writing
one is something any feature does -- and the audit feature never called
insertAuditLog itself, so it was exporting a service purely for someone
else to reach across a boundary for. Both moved to shared.

insertAuditLog was copied verbatim rather than retyped. My first attempt
did retype it and dropped the Content-Profile header and the schema_name
field, neither of which is obvious from the call site. Its tests moved
with it, also verbatim, so the move cannot quietly lose a case.

The remaining ten are the ones that need a design decision rather than an
import fix -- assigned-orders and received-orders, cart and products, and
products and seller-admins each import each other, and a cycle at feature
level usually means one feature wearing two names.

lint 0, typecheck 12/12, admin 531 and payments 554 passing.
