+++ apps/admin/src/features/audit/application/hooks/useAuditLog.ts
@@ -1,4 +1,4 @@
-import { useQuery, useMutation } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 
 import {
   AUDIT_QUERY_KEY,
@@ -8,9 +8,7 @@ import type { AuditFilters } from "@/features/audit/domain/types";
 import {
   fetchAuditLog,
   fetchAuditTableNames,
-  insertAuditLog,
 } from "@/features/audit/infrastructure/auditQueries";
-import { useSupabase } from "@/shared/application/hooks/useSupabase";
 
 /** Keep data fresh for 30s so back-navigation shows cached results instantly */
 const STALE_TIME_MS = 30_000;
@@ -41,16 +39,5 @@ export function useAuditTableNames() {
     queryKey: [TABLE_NAMES_KEY],
     queryFn: () => fetchAuditTableNames(),
     staleTime: STALE_TIME_MS,
-  });
-}
-
-export function useLogExport() {
-  const supabase = useSupabase();
-  return useMutation({
-    mutationFn: async (params: { table: string; count: number }) => {
-      await insertAuditLog(supabase, "EXPORT", params.table, {
-        exported_count: params.count,
-      });
-    },
   });
 }
