+++ apps/admin/tests/auditQueries.test.ts
@@ -209,73 +209,3 @@ describe("fetchAuditLog — branch coverage", () => {
     expect(capturedUrl).not.toContain("action_type");
   });
 });
-
-describe("insertAuditLog", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    vi.stubEnv("NEXT_PUBLIC_SUPABASE_ANON_KEY", "test-anon-key");
-  });
-
-  afterEach(() => {
-    vi.unstubAllEnvs();
-  });
-
-  // insertAuditLog still takes a Supabase client — it needs it to resolve
-  // the caller's local user_profiles.id via getCurrentUserId (mocked above),
-  // separately from the Clerk bearer token used for the raw fetch() itself.
-  const fakeSupabase = {} as Parameters<
-    typeof import("@/features/audit/infrastructure/auditQueries").insertAuditLog
-  >[0];
-
-  it("POSTs to logged_actions and resolves on success", async () => {
-    mockGetSupabaseAccessToken.mockResolvedValue("my-token");
-    mockGetCurrentUserId.mockResolvedValue("user-123");
-
-    let capturedRequest: Request | null = null;
-    server.use(
-      http.post(`${SUPABASE_URL}/rest/v1/logged_actions`, ({ request }) => {
-        capturedRequest = request;
-        return new HttpResponse(null, { status: 201 });
-      }),
-    );
-
-    const { insertAuditLog } =
-      await import("@/features/audit/infrastructure/auditQueries");
-    await insertAuditLog(fakeSupabase, "INSERT", "products", { id: "1" });
-
-    expect(capturedRequest).not.toBeNull();
-    const body = await (capturedRequest as unknown as Request).json();
-    expect(body.action_type).toBe("INSERT");
-    expect(body.table_name).toBe("products");
-    expect(body.row_data).toEqual({ id: "1" });
-    expect(body.user_id).toBe("user-123");
-  });
-
-  it("throws Unauthenticated when there is no Clerk session token", async () => {
-    mockGetSupabaseAccessToken.mockResolvedValue(null);
-
-    const { insertAuditLog } =
-      await import("@/features/audit/infrastructure/auditQueries");
-    await expect(
-      insertAuditLog(fakeSupabase, "INSERT", "products"),
-    ).rejects.toThrow("Unauthenticated");
-  });
-
-  it("throws when the POST response is not ok", async () => {
-    mockGetSupabaseAccessToken.mockResolvedValue("my-token");
-    mockGetCurrentUserId.mockResolvedValue("user-123");
-
-    server.use(
-      http.post(
-        `${SUPABASE_URL}/rest/v1/logged_actions`,
-        () => new HttpResponse(null, { status: 500 }),
-      ),
-    );
-
-    const { insertAuditLog } =
-      await import("@/features/audit/infrastructure/auditQueries");
-    await expect(
-      insertAuditLog(fakeSupabase, "INSERT", "products"),
-    ).rejects.toThrow("Audit insert failed: 500");
-  });
-});
