+++ apps/payments/src/features/assigned-orders/domain/constants.ts
@@ -1,5 +1,3 @@
-export const ASSIGNED_ORDERS_QUERY_KEY = "assigned-orders";
-
 /** Stale time for assigned orders query (30 seconds) */
 export const ASSIGNED_ORDERS_STALE_TIME_MS = 30_000;
