+++ apps/payments/src/features/payment-methods/application/hooks/usePaymentMethods.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "@tanstack/react-query";
 import { useSupabase } from "shared";
 
-import { PAYMENT_METHODS_LIST_QUERY_KEY } from "@/features/payment-methods/domain/constants";
 import { fetchPaymentMethods } from "@/features/payment-methods/infrastructure/paymentMethodQueries";
+import { PAYMENT_METHODS_LIST_QUERY_KEY } from "@/shared/domain/queryKeys";
 
 /** Fetch the seller's configured payment methods (requires sellerId) */
 export function usePaymentMethods(sellerId: string) {
