+++ apps/payments/src/features/received-orders/application/hooks/useReceivedOrders.ts
@@ -3,11 +3,9 @@
 import { useQuery } from "@tanstack/react-query";
 import { useSupabase } from "shared";
 
-import {
-  RECEIVED_ORDERS_QUERY_KEY,
-  RECEIVED_ORDERS_STALE_TIME_MS,
-} from "@/features/received-orders/domain/constants";
+import { RECEIVED_ORDERS_STALE_TIME_MS } from "@/features/received-orders/domain/constants";
 import { fetchReceivedOrders } from "@/features/received-orders/infrastructure/receivedOrderQueries";
+import { RECEIVED_ORDERS_QUERY_KEY } from "@/shared/domain/queryKeys";
 
 export function useReceivedOrders(filter?: string) {
   const supabase = useSupabase();
