+++ apps/payments/src/features/reports/application/hooks/useDelegatedReports.ts
@@ -4,10 +4,10 @@ import { useQuery } from "@tanstack/react-query";
 import { useQueryStates } from "nuqs";
 import { useSupabase } from "shared";
 
-import { DELEGATED_REPORTS_QUERY_KEY } from "@/features/reports/domain/constants";
 import { sellerReportsSearchParams } from "@/features/reports/domain/searchParams";
 import type { SellerReportFilters } from "@/features/reports/domain/types";
 import { fetchDelegatedReportOrders } from "@/features/reports/infrastructure/delegatedReportsApi";
+import { DELEGATED_REPORTS_QUERY_KEY } from "@/shared/domain/queryKeys";
 
 const STALE_TIME_MS = 30_000;
