+++ apps/payments/src/features/reports/domain/constants.ts
@@ -1,5 +1,1 @@
 export { ORDER_STATUS_LIST } from "shared";
-
-export const SELLER_REPORTS_QUERY_KEY = "seller-reports-orders";
-
-export const DELEGATED_REPORTS_QUERY_KEY = "delegated-reports";
