+++ apps/payments/tests/orders-constants.test.ts
@@ -3,10 +3,10 @@ import { describe, it, expect } from "vitest";
 import {
   ACCEPTED_RECEIPT_TYPES,
   MAX_RECEIPT_SIZE_BYTES,
-  MY_ORDERS_QUERY_KEY,
   ORDERS_STALE_TIME_MS,
   RECEIPTS_BUCKET,
 } from "@/features/orders/domain/constants";
+import { MY_ORDERS_QUERY_KEY } from "@/shared/domain/queryKeys";
 
 describe("orders domain constants", () => {
   it("MY_ORDERS_QUERY_KEY is defined", () => {
