+++ apps/payments/tests/payment-methods-constants.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 
-import { PAYMENT_METHODS_QUERY_KEY } from "@/features/payment-methods/domain/constants";
+import { PAYMENT_METHODS_QUERY_KEY } from "@/shared/domain/queryKeys";
 
 describe("payment-methods domain constants", () => {
   it("defines PAYMENT_METHODS_QUERY_KEY", () => {
